Extract dashboard component
