+++ client/src/pages/LandingPage.js
@@ -1,273 +1,35 @@
 import * as React from "react";
-import { styled, createTheme, ThemeProvider } from "@mui/material/styles";
-import CssBaseline from "@mui/material/CssBaseline";
-import MuiDrawer from "@mui/material/Drawer";
-import Box from "@mui/material/Box";
-import MuiAppBar from "@mui/material/AppBar";
-import Toolbar from "@mui/material/Toolbar";
-import List from "@mui/material/List";
-import Typography from "@mui/material/Typography";
-import Divider from "@mui/material/Divider";
-import IconButton from "@mui/material/IconButton";
-import Badge from "@mui/material/Badge";
-import Container from "@mui/material/Container";
-import Grid from "@mui/material/Grid";
-import Paper from "@mui/material/Paper";
-import Link from "@mui/material/Link";
-import MenuIcon from "@mui/icons-material/Menu";
-import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import NotificationsIcon from "@mui/icons-material/Notifications";
-import { useQuery } from '@apollo/client';
+import { Link } from "react-router-dom";
 
-import { mainMenuItems, secondaryMenuItems } from "../components/LeftNavMenuItems";
-import Subscriptions from "../components/Subscriptions";
-//import Chart from "./Chart";
-//import Deposits from "./Deposits";
-
-import Auth from "../utils/auth";
-import { QUERY_ME_BASIC } from "../utils/queries";
-
-function Copyright(props) {
-  return (
-    <Typography
-      variant="body2"
-      color="text.secondary"
-      align="center"
-      {...props}
-    >
-      {"Copyright © "}
-      <Link color="inherit" href="https://mui.com/">
-        Video Streaming Tracker
-      </Link>{" "}
-      {new Date().getFullYear()}
-      {"."}
-    </Typography>
-  );
-}
-
-const drawerWidth = 240;
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== 'open',
-})(({ theme, open }) => ({
-  zIndex: theme.zIndex.drawer + 1,
-  transition: theme.transitions.create(['width', 'margin'], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
-const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
-  ({ theme, open }) => ({
-    '& .MuiDrawer-paper': {
-      position: 'relative',
-      whiteSpace: 'nowrap',
-      width: drawerWidth,
-      transition: theme.transitions.create('width', {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.enteringScreen,
-      }),
-      boxSizing: 'border-box',
-      ...(!open && {
-        overflowX: 'hidden',
-        transition: theme.transitions.create('width', {
-          easing: theme.transitions.easing.sharp,
-          duration: theme.transitions.duration.leavingScreen,
-        }),
-        width: theme.spacing(7),
-        [theme.breakpoints.up('sm')]: {
-          width: theme.spacing(9),
-        },
-      }),
-    },
-  }),
-);
-
-const mdTheme = createTheme();
-
-function DashboardContent() {
-  const [open, setOpen] = React.useState(true);
-  const toggleDrawer = () => {
-    setOpen(!open);
-  };
-
-  const { data: userData } = useQuery(QUERY_ME_BASIC);
-  const loggedIn = Auth.loggedIn();
-
-  return (
-    <ThemeProvider theme={mdTheme}>
-      <Box sx={{ display: "flex" }}>
-        <CssBaseline />
-
-        <AppBar position="absolute" open={open}>
-          <Toolbar
-            sx={{
-              pr: "24px", // keep right padding when drawer closed
-            }}
-          >
-            <IconButton
-              edge="start"
-              color="inherit"
-              aria-label="open drawer"
-              onClick={toggleDrawer}
-              sx={{
-                marginRight: "36px",
-                ...(open && { display: "none" }),
-              }}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography
-              component="h1"
-              variant="h6"
-              color="inherit"
-              noWrap
-              sx={{ flexGrow: 1 }}
-            >
-              Dashboard
-            </Typography>
-            <IconButton color="inherit">
-              <Badge badgeContent={4} color="secondary">
-                <NotificationsIcon />
-              </Badge>
-            </IconButton>
-          </Toolbar>
-        </AppBar>
-
-        <Drawer variant="permanent" open={open}>
-          <Toolbar
-            sx={{
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "flex-end",
-              px: [1],
-            }}
-          >
-            <IconButton onClick={toggleDrawer}>
-              <ChevronLeftIcon />
-            </IconButton>
-          </Toolbar>
-          <Divider />
-          <List component="nav">
-            {mainMenuItems}
-            <Divider sx={{ my: 1 }} />
-            {secondaryMenuItems}
-          </List>
-        </Drawer>
-
-        <Box
-          component="main"
-          sx={{
-            backgroundColor: (theme) =>
-              theme.palette.mode === "light"
-                ? theme.palette.grey[100]
-                : theme.palette.grey[900],
-            flexGrow: 1,
-            height: "100vh",
-            overflow: "auto",
-          }}
-        >
-          <Toolbar />
-          <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-
-            <Grid container spacing={3}>
-
-              {/* List all the user's Subscriptions */}
-
-              {/*
-              {loggedIn && userData ? (
-              */}
-              <Grid item xs={12}>
-                <Paper sx={{ p: 2, display: "flex", flexDirection: "column" }}>
-                  <Subscriptions />
-                </Paper>
-              </Grid>
-              {/*
-              ) : null}
-              */}
-
-              {/* Chart */}
-              {/*
-              <Grid item xs={12} md={8} lg={9}>
-                <Paper
-                  sx={{
-                    p: 2,
-                    display: "flex",
-                    flexDirection: "column",
-                    height: 240,
-                  }}
-                >
-                  <Chart />
-                </Paper>
-              </Grid>
-              */}
-
-              {/* Recent Deposits */}
-              {/* 
-              <Grid item xs={12} md={4} lg={3}>
-                <Paper
-                  sx={{
-                    p: 2,
-                    display: "flex",
-                    flexDirection: "column",
-                    height: 240,
-                  }}
-                >
-                  <Deposits />
-                </Paper>
-              </Grid>
-              */}
-            </Grid>
-
-            <Copyright sx={{ pt: 4 }} />
-          </Container>
-
-        </Box>
-  
-
-      </Box>
-    </ThemeProvider>
-  );
-}
-
-export default function LandingPage() {
-  return <DashboardContent />;
-}
-
-/*
 function LandingPage() {
+  return (
+    <div className="big-image">
+      <div className="overlay">
+        <div className="landing-nav">
+          <Link to="/" className="landing-nav-title">
+            <h1>VideoSubTrack</h1>
+          </Link>
 
-    return (
-        <div className='big-image'>
-            <div className="overlay">
-                <div className='landing-nav'>
-                    <Link to="/" className="landing-nav-title">
-                        <h1>VideoSubTrack</h1>
-                    </Link>
-
-                    <nav className='landing-navbar'>
-                        <>
-                            <Link to='/login' className='landing-list-item'>Login</Link>
-                            <Link to='/signup' className='landing-list-item'>Signup</Link>
-                        </>
-                    </nav>
-                </div>
-                <div>
-                    <p className="hero-text">Video Streaming Subscriptions Tracker </p>
-                    <p className="hero-subtext">Your handy dandy Video Streaming Subscription Ready-Reckoner!</p>
-                </div>
-            </div>
+          <nav className="landing-navbar">
+            <>
+              <Link to="/login" className="landing-list-item">
+                Login
+              </Link>
+              <Link to="/signup" className="landing-list-item">
+                Signup
+              </Link>
+            </>
+          </nav>
         </div>
-    );
+        <div>
+          <p className="hero-text">Video Streaming Subscriptions Tracker </p>
+          <p className="hero-subtext">
+            Your handy dandy Video Streaming Subscription Ready-Reckoner!
+          </p>
+        </div>
+      </div>
+    </div>
+  );
 }
 
 export default LandingPage;
-*/
